refactor(plugin-shape-generator): move LaTeX stuff into separate modules
