+++ src/plugins/shape-generator/index.ts
@@ -2,6 +2,7 @@ import { PluginController } from "#plugins/PluginController.js";
 import { ComputeEngine } from "@cortex-js/compute-engine";
 import { computeEngineLatexToDesmosLatex } from "compute-engine-to-desmos-latex";
 import { ConfigItem } from "..";
+import { ellipseGeneratorExpressions, ellipseLatex } from "./latex/ellipse";
 
 export default class ShapeGenerator extends PluginController<{
   simplifyEquations: boolean;
@@ -59,14 +60,17 @@ export default class ShapeGenerator extends PluginController<{
   }
 
   afterDisable() {
+    // Prevent the Add ellipse button from being added to the Add expression dropdown
     const addExpressionBtn = getAddExpressionButton();
     addExpressionBtn.removeEventListener(
       "click",
       this._addExpressionBtnClickHandler!
     );
 
+    // Remove stylesheet
     document.getElementById("shape-generator-styles")!.remove();
 
+    // Remove generated expressions
     this.cleanupExpressions();
   }
 
@@ -136,9 +140,6 @@ function addExpressionBtnClickHandler(this: ShapeGenerator) {
       label: "ellipse",
       disabled: this.isEditingShape,
       handler: () => {
-        // Close the dropdown
-        addExpressionBtn.dispatchEvent(new CustomEvent("dcg-tap"));
-
         this.isEditingShape = true;
 
         this.calc.observeEvent("change.shapeGenerator", () => {
@@ -219,103 +220,7 @@ function addExpressionBtnClickHandler(this: ShapeGenerator) {
           this.calc.unobserveEvent("change.shapeGenerator");
         });
 
-        this.calc.setExpressions([
-          {
-            id: "shape-generator-ellipse",
-            type: "expression",
-            latex: ellipseLatex(
-              "x_{ellipseGenerator}",
-              "y_{ellipseGenerator}",
-              "r_{xEllipseGenerator}",
-              "r_{yEllipseGenerator}",
-              "A_{ellipseGenerator}"
-            ),
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-position-point",
-            type: "expression",
-            latex: "\\left(x_{ellipseGenerator},y_{ellipseGenerator}\\right)",
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-position-x",
-            type: "expression",
-            latex: "x_{ellipseGenerator}=0",
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-position-y",
-            type: "expression",
-            latex: "y_{ellipseGenerator}=0",
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-radius-x",
-            type: "expression",
-            latex: "r_{xEllipseGenerator}=1",
-            sliderBounds: { min: 0, max: "", step: "" },
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-radius-x-point",
-            type: "expression",
-            latex: `\\left(x_{ellipseGenerator}-r_{xEllipseGenerator}\\cos A_{ellipseGenerator},y_{ellipseGenerator}-r_{xEllipseGenerator}\\sin A_{ellipseGenerator}\\right)`,
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-radius-y",
-            type: "expression",
-            latex: "r_{yEllipseGenerator}=1",
-            sliderBounds: { min: 0, max: "", step: "" },
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-radius-y-point-helper",
-            type: "expression",
-            latex:
-              "f_{ellipseGeneratorRadiusYX}\\left(R_{yEllipseGenerator}\\right)=x_{ellipseGenerator}+R_{yEllipseGenerator}\\cos\\left(A_{ellipseGenerator}+\\frac{\\pi}{2}\\right)",
-            hidden: true,
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-radius-y-point",
-            type: "expression",
-            latex: `\\left(f_{ellipseGeneratorRadiusYX}\\left(r_{yEllipseGenerator}\\right),y_{ellipseGenerator}+r_{yEllipseGenerator}\\sin\\left(A_{ellipseGenerator}+\\frac{\\pi}{2}\\right)\\right)`,
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-angle",
-            type: "expression",
-            latex: "A_{ellipseGenerator}=0",
-            sliderBounds: { min: 0, max: "\\tau", step: "" },
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-angle-point-helper-x",
-            type: "expression",
-            latex:
-              "f_{ellipseGeneratorAngleX}\\left(a_{ellipseGenerator}\\right)=x_{ellipseGenerator}+r_{xEllipseGenerator}\\cos a_{ellipseGenerator}",
-            hidden: true,
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-angle-point-helper-y",
-            type: "expression",
-            latex:
-              "f_{ellipseGeneratorAngleY}\\left(a_{ellipseGenerator}\\right)=y_{ellipseGenerator}+r_{xEllipseGenerator}\\sin a_{ellipseGenerator}",
-            hidden: true,
-            color: "",
-          },
-          {
-            id: "shape-generator-ellipse-angle-point",
-            type: "expression",
-            latex:
-              "\\left(f_{ellipseGeneratorAngleX}\\left(A_{ellipseGenerator}\\right),f_{ellipseGeneratorAngleY}\\left(A_{ellipseGenerator}\\right)\\right)",
-            dragMode: "X",
-            color: "",
-          },
-        ]);
+        this.calc.setExpressions(ellipseGeneratorExpressions);
       },
     },
   ];
@@ -342,7 +247,13 @@ function addExpressionBtnClickHandler(this: ShapeGenerator) {
       newDropdownItemElm.ariaDisabled = "true";
       newDropdownItemElm.classList.add("dcg-disabled");
     } else {
-      newDropdownItemElm.addEventListener("click", newDropdownItem.handler);
+      newDropdownItemElm.addEventListener("click", () => {
+        // Close the dropdown
+        addExpressionBtn.dispatchEvent(new CustomEvent("dcg-tap"));
+
+        // Run the handler for this dropdown item
+        newDropdownItem.handler();
+      });
     }
   }
 }
@@ -367,26 +278,4 @@ function getAddExpressionDropdown() {
   }
 
   return dropdown as HTMLDivElement;
-}
-
-// From
-// https://github.com/lafkpages/desmos-experiments/blob/734279478aaf8162c725b55cd6770480b31bb85f/src/routes/ellipse/latex.ts
-export function ellipseLatex(
-  x: string | number,
-  y: string | number,
-  rx: string | number,
-  ry: string | number,
-  angle: string | number
-) {
-  x = formatArg(x);
-  y = formatArg(y);
-  rx = formatArg(rx);
-  ry = formatArg(ry);
-  angle = formatArg(angle);
-
-  return `\\frac{\\left(\\left(x-${x}\\right)\\cos\\left(${angle}\\right)+\\left(y-${y}\\right)\\sin\\left(${angle}\\right)\\right)^{2}}{${rx}^{2}}+\\frac{\\left(\\left(x-${x}\\right)\\sin\\left(${angle}\\right)-\\left(y-${y}\\right)\\cos\\left(${angle}\\right)\\right)^{2}}{${ry}^{2}}<1`;
-}
-
-export function formatArg(arg: string | number) {
-  return typeof arg === "number" ? `\\left(${arg}\\right)` : arg;
 }
